requirements-dev.txt
pip-chill
python-dotenv==0.19.2
flask-migrate
pygount